. PIC-2290 . Moved enums to separate package. . Added HearingEventType enum
