+++ src/main/java/uk/gov/justice/probation/courtcasematcher/restclient/model/courtcaseservice/CCSOffenderMatch.java
@@ -5,7 +5,7 @@ import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Data;
 import lombok.NoArgsConstructor;
-import uk.gov.justice.probation.courtcasematcher.model.domain.MatchType;
+import uk.gov.justice.probation.courtcasematcher.model.type.MatchType;
 import uk.gov.justice.probation.courtcasematcher.model.domain.OffenderMatch;
 
 import javax.validation.Valid;
