+++ src/test/java/uk/gov/justice/probation/courtcasematcher/messaging/model/commonplatform/CPDefendantTest.java
@@ -1,7 +1,7 @@
 package uk.gov.justice.probation.courtcasematcher.messaging.model.commonplatform;
 
 import org.junit.jupiter.api.Test;
-import uk.gov.justice.probation.courtcasematcher.model.domain.DefendantType;
+import uk.gov.justice.probation.courtcasematcher.model.type.DefendantType;
 import uk.gov.justice.probation.courtcasematcher.model.domain.PhoneNumber;
 
 import java.time.LocalDate;
