+++ src/test/java/uk/gov/justice/probation/courtcasematcher/model/domain/DefendantTest.java
@@ -1,6 +1,7 @@
 package uk.gov.justice.probation.courtcasematcher.model.domain;
 
 import org.junit.jupiter.api.Test;
+import uk.gov.justice.probation.courtcasematcher.model.type.DefendantType;
 
 import static org.assertj.core.api.AssertionsForClassTypes.assertThat;
