+++ src/test/java/uk/gov/justice/probation/courtcasematcher/model/domain/DefendantTypeTest.java
@@ -4,6 +4,7 @@ import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.ValueSource;
+import uk.gov.justice.probation.courtcasematcher.model.type.DefendantType;
 
 import static org.assertj.core.api.Assertions.assertThat;
